Time: 4 ms (15.41%) | Memory: 59.2 MB (37.98%) - LeetSync

+++ 1445-number-of-sub-arrays-of-size-k-and-average-greater-than-or-equal-to-threshold/number-of-sub-arrays-of-size-k-and-average-greater-than-or-equal-to-threshold.cpp
@@ -5,12 +5,12 @@ public:
         for(int i = 0;i<k;i++){
             current_sum+=arr[i];
         }
-        if(current_sum/k>=(double)threshold){
+        if(current_sum/k>=threshold){
             count++;
         }
         for(int i = k;i<arr.size();i++){
             current_sum += arr[i] - arr[i-k];
-            if(current_sum/k>=(double)threshold){
+            if(current_sum/k>=threshold){
                 count++;
             }
         }
